+++ daemon/daemon_windows.go
@@ -16,8 +16,8 @@ import (
 	networktypes "github.com/docker/docker/api/types/network"
 	"github.com/docker/docker/daemon/config"
 	"github.com/docker/docker/daemon/container"
+	"github.com/docker/docker/daemon/internal/libcontainerd/local"
 	"github.com/docker/docker/daemon/network"
-	"github.com/docker/docker/libcontainerd/local"
 	"github.com/docker/docker/libcontainerd/remote"
 	"github.com/docker/docker/libnetwork"
 	nwconfig "github.com/docker/docker/libnetwork/config"
